+++ run-tests.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
-lein with-profile last-clojure,oldest-datomic test
-lein with-profile last-clojure,last-datomic test
+lein with-profile test,last-clojure,oldest-datomic test
+lein with-profile test,last-clojure,last-datomic test
 
-lein with-profile clojure7,oldest-datomic test
+lein with-profile test,clojure7,oldest-datomic test
 
-lein with-profile clojure6,oldest-datomic test
+lein with-profile test,clojure6,oldest-datomic test